Src/BScript/Commands/ExpressionCommand.cs
namespace BScript.Commands
{
    using System;
    using System.Collections.Generic;
    using System.Linq;
    using System.Text;
    using BScript.Expressions;

    public class ExpressionCommand
    {
        private IExpression expression;

        public ExpressionCommand(IExpression expression)
        {
            this.expression = expression;
        }

        public IExpression Expression { get { return this.expression; } }

        public void Execute(Context context)
        {
            this.expression.Evaluate(context);
        }
    }
}
